requirements.txt
bokeh==2.2.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
line-bot-sdk==1.18.0
MarkupSafe==1.1.1
matplotlib==3.3.3
mysql-connector-python==8.0.22
numpy==1.19.5
oauthlib==3.1.0
packaging==20.8
pandas==1.2.0
Pillow==8.1.0
protobuf==3.14.0
psycopg2==2.8.6
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
SQLAlchemy==1.3.22
tornado==6.1
tweepy==3.10.0
typing-extensions==3.7.4.3
tzdata==2020.5
urllib3==1.26.2
Werkzeug==1.0.1
gunicorn==20.0.4
